try to fix CI
a newer version of babel seems to complain when being passed superfluous
args. by default, the test.sh script passes all args to the
compilation-command (which is executed if the compiled files don't exist
or outdated). most of the time, the added arguments are meant for the
actual test execution, though. therefore, I think we can remove the
additional arg passing to prepare. for the case, where the entire
test.sh script was called with "prepare", the arguments are still
passed.