fine tune influence of gpu factor
1) The old behavior effectively chose 2048 buckets as a default (for
medium) because this requires only one data texture to be compatible
with the actual requirement of 1536 buckets. Therefore, medium was
changed to require 2048 buckets instead.
2) If a bucket limit was satisfied with a "too large" texture (to avoid
multiple textures), this also meant that wk tried to use the full
capacity. I changed the code so that the current setting of the GPU
factor controls the exact limit of the bucket picker. This means that
changing the hardware utilization can have an immediate impact on the
rendered quality (given that the allocated textures allow for it).

Also see issue 6897.