Fix triggering of volume annotation materialization.
The required property of this rule was misleading since it is not required if
a bounding box cannot be configured, but in addition seems to be ignored if a
validator is set.
Thus in the validator one needs to make sure to avoid letting the validation
fail if a bounding box cannot be configured.

Also improve validation messages for bounding box selection form item.