Fix flood fill tool for coarse mags (#8382)
* Fix flood fill tool for coarse mags.

The 2D flood fill tool did not work properly when flood filling in a mag
other than the finest one if the "third-dimension" coordinate was not
divisible by the respective mag. This resulted in early-out flood fills
only filling a single voxel.
Also, this commit adjusts the flood fill maximum bounding size limits if
the segmentation layer's finest mag is not 1. This allows to use the
flood fill tool in coarser mags if the segmentation layer is properly
mag-restricted.

* update changelog

* Update frontend/javascripts/oxalis/model/sagas/volume/floodfill_saga.tsx

Co-authored-by: Philipp Otto <philippotto@users.noreply.github.com>

* Apply naming feedback

* Add linting rule for unused imports and fix 10 occurrences

* Fix 3 new unused imports after master merge

---------

Co-authored-by: Philipp Otto <philippotto@users.noreply.github.com>