Fix health check (version mismatch check for tracingstore) (#8683)
Ignore whitespace changes while reviewing.

The PR corrects the type definitions for the buildinfo and healthcheck
related requests and therefore fixes a problem with a healthcheck for
dedicated tracingstores.

### URL of deployed dev instance (used for testing):
- https://___.webknossos.xyz

### Steps to test:
- I don't really know of a good way to test this on a dev instance. The
healthcheck for dedicated tracingstores failed anyway, so it shouldn't
get worse now. Also the improved type definitions are a good-enough
guarantee in my opinion.

### Issues:
- fixes #6176 

------
- [X] Updated
[changelog](../blob/master/CHANGELOG.unreleased.md#unreleased)