Fix nightly, improve browser feature check and misc other improvements (#8702)
- wkorg screenshot test failed because of changed view configuration
defaults --> the wkorg test now overwrites the defaults on each run.
this should be in our best interest, anyway, because it ensures good
defaults. the downside is that we have to adapt the code if we want to
tune the default.
- flatMap was not supported by the browser used on browserstack. I fixed
this by specifying a newer osx version.
- flatMap on generators is now checked in the browser feature tests so
that users are properly alerted about a necessary browser update
- errors in the model initialization are detected in the screenshot
tests now so that there is a better error message (and an eager quit on
errors). until now, the test just waited until the timeout was exceeded.
- there were two problems with float32 datasets:
- the "auto clipping" feature did not work with infinite values (which
exist in the float32 test dataset)
- osx ventura did have shader problems with wide float32 ranges. using
sequoia fixes this.

### URL of deployed dev instance (used for testing):
- https://___.webknossos.xyz

### Steps to test:
- see this run
https://github.com/scalableminds/webknossos/actions/runs/15821624973

### Issues:
- fixes failing nightly

------
(Please delete unneeded items, merge only when none are left open)
- [X] Added changelog entry (create a `$PR_NUMBER.md` file in
`unreleased_changes` or use `./tools/create-changelog-entry.py`)