Improve error reporting when sagas crash (#8712)
Errors within the sagas are critical. They are reported currently, but
if a simple failing network request brings down the entire sagas (most
have a retry mechanism but some don't), the current error reporting only
shows the failing network request (which doesn't really show the
severity).

### URL of deployed dev instance (used for testing):
- https://___.webknossos.xyz

### Steps to test:
- I added a `throw new Error("bla")` to the brush saga. you can see the
report for it here:
https://scm.airbrake.io/projects/95438/groups/4097474921126226079?tab=notice-detail
- no need to test further imo

### Issues:
- improves error reporting