Add HTTP API versioning for DataStore and TracingStore (#8705)
Introducing versioning for the HTTP APIs of datastore and tracingstore,
analogous to the existing one on the wk side. That means that e.g.
http://localhost:9000/data/v9/health should now reply the same as
http://localhost:9000/data/health – and we now have the option to
support legacy variants of datastore and tracingstore routes (for a
limited time).

As discussed in
[slack](https://scm.slack.com/archives/C5AKLAV0B/p1750402140167989) this
is now designed such that the versions of wk, datastore and tracingstore
have to be identical.

That means than when we introduce a new version in one of them, we have
to create version redirects for the other two as well.

I tried to unify the formatting of the routes files a little and moved
the collections package in util to unify package structure there.

### URL of deployed dev instance (used for testing):
- https://dstsversioning.webknossos.xyz

### Steps to test:
- Trace some, should work normally
- Try some routes with explicit version. Should work as expected (5 to 9
should be equal to latest, newer and older than that should be rejected)

### Issues:
- fixes #8235 
- Corresponding libs issue (blocked by this PR):
https://github.com/scalableminds/webknossos-libs/issues/1327

------
- [x] Added changelog entry (create a `$PR_NUMBER.md` file in
`unreleased_changes` or use `./tools/create-changelog-entry.py`)
- [x] Considered [common edge
cases](../blob/master/.github/common_edge_cases.md)
- [x] Needs datastore update after deployment