Fix duplicating compound volume annotations (#8743)
The tracingstore needs to be able to get the datasource for an
annotation in order to duplicate it. This is for fallback layer handling
and has been introduced in its current form in #8405

The wk side answers this with either postgres (stored annotations) or a
special temporary store (for temporary annotations). This PR includes
compound annotations in this handling for temporary annotations.

### URL of deployed dev instance (used for testing):
- https://fixduplicatecompound.webknossos.xyz

### Steps to test:
- Create volume task type, project, tasks
- Trace some
- View compound annotation
- hit duplicate
- Should result in a usable explorative annotation.

### Issues:
- fixes #8736

------
- [x] Added changelog entry (create a `$PR_NUMBER.md` file in
`unreleased_changes` or use `./tools/create-changelog-entry.py`)
- [x] Considered [common edge
cases](../blob/master/.github/common_edge_cases.md)
- [x] Needs datastore update after deployment