Truncate action log to avoid json string decoding overflow  (#8757)
This PR introduces a new truncate url param to the update actions log
route. If set to true (default should be false), the backend sends the a
max of 1000 update actions per update group. The goal is to reduce the
max possible string length of the json encoded update actions to not hit
the limit of the frontend. See
[#8594](https://github.com/scalableminds/webknossos/issues/8594) for
more details.

### URL of deployed dev instance (used for testing):
- https://___.webknossos.xyz

### Steps to test:
- I suggest testing locally.
- Open TSAnnotationService and set `maxUpdateActionLogBatchSize = 3` or
so.
- Open an annotation and annotate buckets multiple times with multiple
strokes
- Save
- Open the version restore view. All version numbers should still exist.
When inspecting the network requests, a maximum of 3 update actions per
version should exist.

### TODOs:
- [x] confirm whether that's actually what we want.

### Issues:
- fixes [#8594](https://github.com/scalableminds/webknossos/issues/8594)

------
(Please delete unneeded items, merge only when none are left open)
- [x] Added changelog entry (create a `$PR_NUMBER.md` file in
`unreleased_changes` or use `./tools/create-changelog-entry.py`)

---------

Co-authored-by: Philipp Otto <philippotto@users.noreply.github.com>
Co-authored-by: Florian M <fm3@users.noreply.github.com>