When determining mag realpaths, treat schemeless uris as local (#8804)
Path.of fails for URIs without scheme. We want to assume local files in
this case.

I don’t like that this is encoded with null and we need to check this in
so many places. Refactoring issue for that:
https://github.com/scalableminds/webknossos/issues/8762

### Steps to test:
- In a datasource-properties.json, add a mag with a path that is
absolute and inside of the organization dir.
- Dataset scanning should not report `Failed to determine real paths of
mags of <ds>`. Realpaths should be set correctly in the DB.

------
- [x] Considered [common edge
cases](../blob/master/.github/common_edge_cases.md)
- [x] Needs datastore update after deployment