Fix format in volume annotation download zarr.json (#8800)
Fixes a regression introduced in #7995 – the volume buckets are blosc
compressed, but that was no longer mentioned in the zarr header.

Also, the chunks had the `c.` name prefix, but `chunk_key_encoding` was
already `v2` in the zarr header. Now I actually remove the `c.` name
prefix so that it is actually `v2` compliant.

Also, fixes passing the voxel size to the volume zip download so that it
is correct in the included datasource-properties.json

### Steps to test:
- Create volume annotation, download it in zarr format
- Unpack the zip and the contained data_Volume.zip
- place the result in the binaryData folder, should be readable as
standalone dataset
- reupload the downloaded annotation, should still be uploadable as an
annotation
- Also, upload volume annotation downloaded on the *master* code, should
also still be uploadable as annotation

### Issues:
- fixes #7961 

------
- [x] Added changelog entry (create a `$PR_NUMBER.md` file in
`unreleased_changes` or use `./tools/create-changelog-entry.py`)
- [x] Considered [common edge
cases](../blob/master/.github/common_edge_cases.md)

---------

Co-authored-by: MichaelBuessemeyer <39529669+MichaelBuessemeyer@users.noreply.github.com>