Fix viewing zoomed out datasets for which rotations are configured transformation via dataset settings (#8819)
The fix works by disabling the vertex position optimization code for
datasets where all layers are transformed.
Previously a dataset where all layers had transforms had
`representativeMagForVertexAlignment=[1,1,1]` regardless of the zooming
of the user. This caused the vertices of the plane to be bucket aligned
to mag 1 buckets by the optimization code. The calculated bucket
positions were then disregarded in the fragment shader but lead to the
"blank" areas. Because when zoomed out, there were not enough vertices
to cover all buckets in mag 1 regarding the viewed area leaving gaps in
what was rendered.
This should now be fixed.

### URL of deployed dev instance (used for testing):
- https://___.webknossos.xyz

### Steps to test:
- Open DS settings for a dataset and configure a default rotation of
e.g. 180° around x
- View the dataset and zoom out as far as possible -> everything should
look normal
- Checkout master. View the DS again -> now not all of the screen should
be a "rendered area" shown by the viewports default color being rendered
for parts of the viewport.

### Issues:
- reported here:
https://scm.slack.com/archives/C5AKLAV0B/p1753954379767099

------
(Please delete unneeded items, merge only when none are left open)
- [x] Added changelog entry (create a `$PR_NUMBER.md` file in
`unreleased_changes` or use `./tools/create-changelog-entry.py`)

---------

Co-authored-by: Florian M <fm3@users.noreply.github.com>