Fix positionForSegment route (#8830)
#8708 missed to update the frontend part of the `positionForSegment`
route. I checked this in the changes of #8708. This should fix this.
Furthermore, I did investige some other potential routes and did not
find any other being broken. But I did not a full search, just some
other related routes in `datastore.latest.routes`.

### URL of deployed dev instance (used for testing):
- https://___.webknossos.xyz

### Steps to test:
- Not sure how to test this reliably. I triggered the bug locally by
splitting two segments that had no connecting surface in the agglomerate
mesh. This lead to a crash of wk and when I investigated this, it showed
that the route I fixed here was causing a 404 on the current master.

### Issues:
- self noticed bug

------
(Please delete unneeded items, merge only when none are left open)
- [x] Added changelog entry (create a `$PR_NUMBER.md` file in
`unreleased_changes` or use `./tools/create-changelog-entry.py`)