Frontend Only: Default to treating a mapping as HDF5 in case it is not json (#8880)
This PR makes the frontend treat any mapping as hdf5 which does not
explicitly claim to be json. This helps as there are now also zarr based
mappings which have the same api as hdf5 stored mappings. Thus treating
them equally in the frontend should be fine. Thus, the frontend is more
friendly when not the "correct" type hdf5 is provided.

### URL of deployed dev instance (used for testing):
- https://___.webknossos.xyz

### Steps to test:
- Open
https://fallbacktohdfforagglomappings.webknossos.xyz/datasets/test-agglomerate-file-68b1a05d010000a500a96176/view.
It should have agglo mapping 70 active.
- open
https://fallbacktohdfforagglomappings.webknossos.xyz/datasets/test-agglomerate-file-68b1a05d010000a500a96176/view#%7B%22position%22:%5B100,100,75%5D,%22mode%22:%22orthogonal%22,%22zoomStep%22:0.311,%22additionalCoordinates%22:%5B%5D,%22nativelyRenderedLayerName%22:null,%22stateByLayer%22:%7B%22segmentation%22:%7B%22mappingInfo%22:%7B%22mappingName%22:%22agglomerate_view_70%22,%22mappingType%22:%22Zarr3%22%7D,%22isDisabled%22:false,%22color%22:%5B255,255,255%5D,%22isInverted%22:false,%22gammaCorrectionValue%22:1%7D%7D%7D
  - should work anyway as mappingType in the link is `Zarr3`
- The rest is rather hard to test as it rather deeply involved in the
code. So maybe the 2 tests above are enought?

### Issues:
- fixes https://scm.slack.com/archives/C5AKLAV0B/p1756454411418029

------
(Please delete unneeded items, merge only when none are left open)
- [x] Added changelog entry (create a `$PR_NUMBER.md` file in
`unreleased_changes` or use `./tools/create-changelog-entry.py`)

---------

Co-authored-by: Daniel <daniel.werner@scalableminds.com>