Fix multiple Frontend API mesh mesh related functions and apply mesh visibility changed done while still loading the mesh (#8936)
This PR fixes the Frontend mesh visibility setting function of the API.
Some PR in the past added an additional nesting level to the meshes
structure to support additional coordinates. This was not adjusted in
the frontend mesh visiblity api function. The function itself has a
check for whether the mesh is present but was missing this additional
nesting level. This PR fixes this. The correct nesting structure is
shown in:
https://github.com/scalableminds/webknossos/blob/master/frontend/javascripts/viewer/model/reducers/annotation_reducer.ts#L420-L432

While working on this, I thought it would be nice to be able to script
something that does some additional followup actions once the mesh is
loaded. But the mesh loading function does not return a promise which
resolves when the mesh is loaded. Therefore to add generality, I added a
utility function to the frontend api that enables us to wait for a
specific action. E.g. the mesh loading has a specific "finished" action
which could be used in the case I described. Please feel free to tell me
what you think about this little addition :)

### URL of deployed dev instance (used for testing):
- https://___.webknossos.xyz

### Steps to test:
- View a DS with a mesh file
- Select the mesh file and its corresponding agglomerate view if needed
- Locate a segment with its ID
- Insert the mesh id and position in the following script and run it.
```javascript
window.webknossos.apiReady(3).then(async (api) => {
  const segmentIdsAndPositions = [
      [<mesh_id>, [<mesh_pos>]],
  ];
  const sleep = (ms) => new Promise(resolve => setTimeout(resolve, ms));
  for (const [segmentId, position] of segmentIdsAndPositions) {
    api.data.loadPrecomputedMesh(segmentId, position);
    api.data.setMeshVisibility(segmentId, false);
  }
  await sleep(5*1000);
  for (const [segmentId, position] of segmentIdsAndPositions) {
     api.data.setMeshVisibility(segmentId, true);
  }
 await sleep(5*1000);
  for (const [segmentId, position] of segmentIdsAndPositions) {
     api.data.removeMesh(segmentId);
  }
 await sleep(5*1000);
  for (const [segmentId, position] of segmentIdsAndPositions) {
     api.data.loadPrecomputedMesh(segmentId, position);
  }
await sleep(5*1000);
  for (const [segmentId, position] of segmentIdsAndPositions) {
     api.data.resetMeshes();
  }
});
```
- This should not crash (see console). Watch the 3d viewport to observe
correctness.
  1. Load mesh invisible
  2. make visible again
  3. remove mesh
  4. load mesh again (visible)
  5. remove all meshes 

- Optional: Test on master -> shouldn't work -> the `setMeshVisibility`
part should print an error in the console.

### Issues:
- fixes https://scm.slack.com/archives/C3PBDFGC8/p1758101435301069

------
(Please delete unneeded items, merge only when none are left open)
- [x] Added changelog entry (create a `$PR_NUMBER.md` file in
`unreleased_changes` or use `./tools/create-changelog-entry.py`)