Allow configuring custom logout redirect for oidc (#9006)
Adds the option to configure a custom redirect uri after the logout.

This was requested so that an OIDC provider logout can also be
triggered.

Note that this does not work for the “logout everywhere” feature. This
is accepted for the moment, as this would be more complex (not sure how
to tell the oidc provider to terminate all sessions).

Note that this also does not yet implement the other direction (if a
user logs out of another app with the same OIDC provider, wk sessions
are not terminated). This would also be a lot more complex.

### Steps to test:
- Adapt logout redirect address in application.conf (e.g. to
https://example.org)
- Log out in wk, should be redirected.
- When navigating to wk again, should still be logged out.

### Issues:
- fixes #9004

------
- [x] Removed dev-only changes like prints and application.conf edits
- [x] Considered [common edge
cases](../blob/master/.github/common_edge_cases.md)