Remove Datastore config reportUsedStorage.enabled. Set in DB instead (#9024)
This config option overwrote the value in postgres with every datastore
ping. We want to be able to set it in the db permanently.

In fact, we set it to true in the test DB. This will be tested in the
libs tests.

### Steps to test
- Set the flag to true in db. It should stay true even through wk
restarts and 10min datastore ping interval.
 - Set the flag to false, should stay false.
 - CI/e2e-tests should be green

------
- [x] Added changelog entry (create a `$PR_NUMBER.md` file in
`unreleased_changes` or use `./tools/create-changelog-entry.py`)
- [x] Added migration guide entry if applicable (edit the same file as
for the changelog)
- [x] Needs datastore update after deployment