Merge RemoteSourceDescriptorService into DataVaultService (#8980)
While it was nice that the DataVaultService was small and clean, it was
weird that callers needed to use RemoteSourceDescriptorService to get
their vaultpaths. Lately the separation felt artificial.

I also renamed the dataclass RemoteSourceDescriptor to
CredentializedUPath, which is hopefully more descriptive.

### URL of deployed dev instance (used for testing):
- https://refactordatavaultservice.webknossos.xyz

### Steps to test:
- Load some datasets, both local and remote, should show data.
- Explore remote datast with credentials, should work.

------
- [x] Removed dev-only changes like prints and application.conf edits
- [x] Considered [common edge
cases](../blob/master/.github/common_edge_cases.md)
- [x] Needs datastore update after deployment