Disallow importing untyped version of "select" saga effect (#9060)
This PR adds a script disallowing to import `select` from
`typed-redux-saga`. We have our own better-typed variation in
`effect-generators`.
This small script should prevent merging code using this import with
weak typing to avoid having to fix bugs like
https://github.com/scalableminds/webknossos/pull/9057.

I also investigated doing this directly in biome, but it seems like
biome does not support such a thing yet. Might be the case in the
future. But this script should be sufficient for now.

### URL of deployed dev instance (used for testing):
- https://___.webknossos.xyz

### Steps to test:
- CI should be enough

### Issues:
- Discussion:
https://scm.slack.com/archives/C5AKLAV0B/p1762769130483109?thread_ts=1762764001.604969&cid=C5AKLAV0B
------
(Please delete unneeded items, merge only when none are left open)
- No todos. Not user facing