Fixed crash during navigation when the page is not reloaded.
The issue happened when navigating to another page that also showed a TracingLayoutView.
For a brief moment, the new view is already rendered, but the old one was not unmounted
yet, which is why the isWkInitialized flag in the store was still set to true. This led to
a premature initialization of a lot of components although the new model was not fetched
which led to page crashes.