Move schema and evolutions (#9090)
Never really made sense to me why the schema would be in conf, and the
evolution in tools. One more toplevel folder is some added noise, but it
still seems most natural to me. Other suggestions welcome.

I have always been cautious with moving code, lest we lose the full git
blame. This is an experiment to see if using only git mv helps with
that. `git blame` in the command line still works correctly. If we do
lose the github web blame on the schema, I think that’s no terrible
loss. If we don’t, I’d like to do more moves soon.

### URL of deployed dev instance (used for testing):
- https://moveschema.webknossos.xyz

### Steps to test:
- Wk should still start normally, load data.
- CI

------
- [x] Added migration guide entry if applicable (edit the same file as
for the changelog)