Fixes in test db for libs tests (#9096)
The libs tests rely on the l4_sample dataset in the test database.
Adding paths for it here. That should hopefully fix the libs nightly. I
already ran the tests locally against this branch’s image and they
worked, but that was just one python version.

Strangely enough, despite the separate commit to move the mags csv, the
github diff now shows it as deleted+new :/