Refactor datastore/tracingstore access checks (#9112)
These access checks in datastore and tracingstore work by asking wk:
does the user as identified by this token have access to read/write/etc
this resource identified by id xyz.

Previously, the id was wrapped in a `DataSourceId` class (which was a
bit of an abuse, since this was also used for things like tracings,
which are definitely no DataSources). This PR refactors this to just use
a generic `resourceId: Option[String]`.

This comes with a change in `DSLegacyApiController`, which was the last
place that actually used access checks by `DataSourceId`s. Instead, it
now looks up dataset ids, and then does the access check with that.

That change, however, must be done in such a way that it does not become
possible to distinguish between inaccessible and non-existent datasets,
so users can’t leech information about other organizations. That’s why I
added the `withResolvedDatasetId` function to `DSLegacyApiController`.

### URL of deployed dev instance (used for testing):
- https://refactoraccesschecks.webknossos.xyz

### Steps to test:
- Do some normal annotating, should still work
- try accessing other user’s work, which you should not have access to.
Access should be denied.
- by requesting older versions of datastore routes (with
orgaId+datasetDirectoryName) you should not be able to distinguish
between datasets you don’t have access to, and datasets that don’t
exist.

### Issues:
- fixes #8239

------
- [x] Removed dev-only changes like prints and application.conf edits
- [x] Considered [common edge
cases](../blob/master/.github/common_edge_cases.md)
- [x] Needs datastore update after deployment

---------

Co-authored-by: MichaelBuessemeyer <39529669+MichaelBuessemeyer@users.noreply.github.com>